backend/pyproject.toml
[project]
name = "troca-uma-ideia"
version = "0.1.0"
description = "Backend do projeto Troca uma Ideia"
authors = [
    { name = "AndreDG88", email = "andre.soares.designer@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django ==4.2.21",
    "djangorestframework >=3.16.0,<4.0.0",
    "psycopg2-binary >=2.9.10,<3.0.0",
    "python-decouple >=3.8,<4.0",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 88
target-version = ['py311']
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "trocaumaideia.settings"
python_files = "test_*.py"




